+++ sdk/businessAccountRequest/index.js
@@ -1,0 +1,10 @@
+const businessAccountRequest = require('./businessAccountRequest.js');
+
+exports.log = require('./log');
+exports.create = businessAccountRequest.create;
+exports.get = businessAccountRequest.get;
+exports.query = businessAccountRequest.query;
+exports.page = businessAccountRequest.page;
+exports.BusinessAccountRequest = businessAccountRequest.BusinessAccountRequest;
+exports.Address = require('./address.js').Address;
+exports.Owner = require('./owner.js').Owner;